style : refactor, move to new file and comment
