+++ crates/crabe_decision/src/utils.rs
@@ -3,3 +3,6 @@ pub use self::utils::*;
 
 pub mod constants;
 pub use self::constants::*;
+
+pub mod navigation;
+pub use self::navigation::*;
